changed artiq experiments around
